// src/2022/02/02-enums.ts
export enum OpponentTurn {
    ROCK = 'A',
    PAPER = 'B',
    SCISSORS = 'C',
}

export enum MyTurn {
    ROCK = 'X',
    PAPER = 'Y',
    SCISSORS = 'Z',
}

export enum HowTurnShouldEnd {
    LOST = 'X',
    DRAW = 'Y',
    WIN = 'Z',
}

export enum RoundPoints {
    LOST = 0,
    DRAW = 3,
    WIN = 6,
}